Adding searching, sorting and pagination capabilities to the tables. Moved the backend scripts to a different table
