+++ scripts/story1.py
