+++ scripts/story2.py
